[[ BytecodeBlocks ]] Add bytecode blocks
This patch adds the ability to write bytecode directly inside 'bytecode'
blocks in LCB:

    bytecode
      ...
    end bytecode

Labels are local to the block and can be specified by using:

    myLabel:

Temporaries are local to the block and can be specified by using:

   variable myTemp

Temporaries, local variables and parameters can be used as register type
arguments to opcodes.

Module variables, defined constants and handler identifiers can be used
as the 'source' parameter for the 'fetch' opcode.

Module variables can be used as the 'dest' parameter for the 'source'
opcode.

Handler identifiers can be used as the 'handler' parameter for the 'invoke'
opcode.

The bytecode information is recorded in a simple array in
script-builder.cpp - including name, opcode index and opcode
parameter format.

The script builder APIs have been thinned down to remove specific
bytecode emission functions. Instead, bytecode is emitted using
EmitBytecodeInModule. The emit module of lc-compile has been
updated to use this.